requirements.txt
# ── Core API ─────────────────────────────────────────
fastapi==0.110.0
uvicorn[standard]==0.27.1

# ── Data & ML ────────────────────────────────────────
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1.post1

# ── Deep Learning ────────────────────────────────────
torch==2.2.1

# ── NLP ──────────────────────────────────────────────
nltk==3.8.1

# ── HTTP Client (for Groq API) ───────────────────────
httpx==0.27.0

# ── Environment Config ───────────────────────────────
python-dotenv==1.0.1

# ── Validation ───────────────────────────────────────
pydantic==2.6.3

# ── Optional (recommended for performance) ───────────
orjson==3.9.15